chore(ui): split compliance results service
