+++ ui/apps/platform/src/services/ComplianceResultsStatsService.ts
@@ -1,0 +1,24 @@
+import axios from 'services/instance';
+
+import { ComplianceCheckStatusCount, complianceV2Url } from './ComplianceCommon';
+
+type ComplianceProfileScanStats = {
+    checkStats: ComplianceCheckStatusCount[];
+    profileName: string;
+    title: string;
+    version: string;
+};
+
+export type ListComplianceProfileScanStatsResponse = {
+    scanStats: ComplianceProfileScanStats[];
+    totalCount: number;
+};
+
+/**
+ * Fetches the scan stats grouped by profile.
+ */
+export function getComplianceProfilesStats(): Promise<ListComplianceProfileScanStatsResponse> {
+    return axios
+        .get<ListComplianceProfileScanStatsResponse>(`${complianceV2Url}/scan/stats/profiles`)
+        .then((response) => response.data);
+}
